+++ backend/baton/src/main/java/touch/baton/domain/tag/query/service/TagQueryService.java
@@ -1,11 +1,12 @@
 package touch.baton.domain.tag.query.service;
 
+import jakarta.annotation.Nullable;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import touch.baton.domain.tag.command.Tag;
 import touch.baton.domain.tag.command.vo.TagReducedName;
-import touch.baton.domain.tag.query.repository.TagQueryRepository;
+import touch.baton.domain.tag.query.repository.TagQuerydslRepository;
 
 import java.util.List;
 
@@ -14,9 +15,9 @@ import java.util.List;
 @Service
 public class TagQueryService {
 
-    private final TagQueryRepository tagQueryRepository;
+    private final TagQuerydslRepository tagQuerydslRepository;
 
-    public List<Tag> readTagsByReducedName(final String tagName) {
-        return tagQueryRepository.readTagsByReducedName(TagReducedName.from(tagName));
+    public List<Tag> readTagsByReducedName(@Nullable final TagReducedName tagReducedName, final int limit) {
+        return tagQuerydslRepository.findByTagReducedName(tagReducedName, limit);
     }
 }
