+++ backend/baton/src/test/java/touch/baton/assure/repository/TestSupporterRunnerPostQueryRepository.java
@@ -4,7 +4,6 @@ import org.springframework.context.annotation.Profile;
 import touch.baton.domain.member.query.repository.SupporterRunnerPostQueryRepository;
 
 @Profile("test")
-
 public interface TestSupporterRunnerPostQueryRepository extends SupporterRunnerPostQueryRepository {
 
     default Long getApplicantCountByRunnerPostId(final Long runnerPostId) {
