+++ backend/baton/src/test/java/touch/baton/domain/tag/command/vo/TagReducedNameTest.java
@@ -1,8 +1,7 @@
-package touch.baton.domain.tag.vo;
+package touch.baton.domain.tag.command.vo;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import touch.baton.domain.tag.command.vo.TagReducedName;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
